Handle presets, remove some deployment stuff
