--- core-ui/src/components/Predefined/Create/Deployments/AdvancedForm.scss
@@ -1,5 +1,0 @@
-.runtime-profile-form {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: 16px;
-}
