+++ core-ui/src/components/Predefined/Create/Deployments/Deployments.create.js
@@ -4,14 +4,7 @@ import { LabelsInput } from 'components/Lambdas/components';
 import { useTranslation } from 'react-i18next';
 import { K8sNameInput, usePost, useNotification } from 'react-shared';
 import { ResourceForm } from './../ResourceForm/ResourceForm';
-import {
-  Button,
-  FormFieldset,
-  FormItem,
-  FormLabel,
-  Checkbox,
-  MessageStrip,
-} from 'fundamental-react';
+import { Button, Checkbox } from 'fundamental-react';
 import * as jp from 'jsonpath';
 import './Deployments.create.scss';
 
@@ -21,6 +14,7 @@ import {
   createPresets,
   createServiceTemplate,
 } from './templates';
+import { Containers } from './Containers';
 
 export function DeploymentsCreate({
   formElementRef,
@@ -129,6 +123,13 @@ export function DeploymentsCreate({
       onCreate={onCreate}
       renderEditor={renderEditor}
       presets={createPresets(namespace, t)}
+      onPresetSelected={({ value }) => {
+        setDeployment(value.deployment);
+        setCreateService(!!value.service);
+        if (value.service) {
+          setService(value.service);
+        }
+      }}
     >
       <ResourceForm.FormField
         required
@@ -178,6 +179,7 @@ export function DeploymentsCreate({
       />
       <ResourceForm.FormField
         advanced
+        className="fd-margin-bottom--sm"
         propertyPath="$.metadata.annotations"
         label={t('common.headers.annotations')}
         input={(value, setValue) => (
@@ -263,150 +265,4 @@ export function DeploymentsCreate({
       </ResourceForm.CollapsibleSection>
     </ResourceForm>
   );
-}
-
-function SingleContainerSection({ container, containers, setContainers }) {
-  const { t } = useTranslation();
-
-  return (
-    <>
-      <ResourceForm.FormField
-        label="Name"
-        value={container.name}
-        setValue={name => {
-          container.name = name;
-          setContainers([...containers]);
-        }}
-        required
-        input={(value, onChange) => (
-          <ResourceForm.Input required value={value} setValue={onChange} />
-        )}
-      />
-      <ResourceForm.FormField
-        label="Image"
-        value={container.image}
-        setValue={image => {
-          container.image = image;
-          setContainers([...containers]);
-        }}
-        required
-        input={(value, onChange) => (
-          <ResourceForm.Input required value={value} setValue={onChange} />
-        )}
-        className="fd-margin-bottom--sm"
-      />
-
-      <ResourceForm.CollapsibleSection
-        title={t('deployments.create-modal.advanced.runtime-profile')}
-        canChangeState={false}
-        defaultOpen
-      >
-        <FormFieldset className="runtime-profile-form">
-          <FormItem>
-            <FormLabel required>
-              {t('deployments.create-modal.advanced.memory-requests')}
-            </FormLabel>
-            <ResourceForm.Input
-              required
-              value={jp.value(container, '$.resources.requests.memory') || ''}
-              setValue={memory => {
-                jp.value(container, '$.resources.requests.memory', memory);
-                setContainers([...containers]);
-              }}
-            />
-          </FormItem>
-          <FormItem>
-            <FormLabel required>
-              {t('deployments.create-modal.advanced.memory-limits')}
-            </FormLabel>
-            <ResourceForm.Input
-              required
-              value={jp.value(container, '$.resources.limits.memory') || ''}
-              setValue={memory => {
-                jp.value(container, '$.resources.limits.memory', memory);
-                setContainers([...containers]);
-              }}
-            />
-          </FormItem>
-        </FormFieldset>
-        <FormFieldset className="runtime-profile-form">
-          <FormItem>
-            <FormLabel required>
-              {t('deployments.create-modal.advanced.cpu-requests')}
-            </FormLabel>
-            <ResourceForm.Input
-              required
-              value={jp.value(container, '$.resources.requests.cpu') || ''}
-              setValue={cpu => {
-                jp.value(container, '$.resources.requests.cpu', cpu);
-                setContainers([...containers]);
-              }}
-            />
-          </FormItem>
-          <FormItem>
-            <FormLabel required>
-              {t('deployments.create-modal.advanced.cpu-limits')}
-            </FormLabel>
-            <ResourceForm.Input
-              required
-              value={jp.value(container, '$.resources.limits.cpu') || ''}
-              setValue={cpu => {
-                jp.value(container, '$.resources.limits.cpu', cpu);
-                setContainers([...containers]);
-              }}
-            />
-          </FormItem>
-        </FormFieldset>
-      </ResourceForm.CollapsibleSection>
-    </>
-  );
-}
-
-function Containers({ value: containers, setValue: setContainers }) {
-  const { t } = useTranslation();
-
-  const removeContainer = index => {
-    setContainers(containers.filter((_, i) => index !== i));
-  };
-
-  containers = containers || [];
-
-  if (!containers.length) {
-    return (
-      <MessageStrip type="warning">
-        At least one container is required.
-      </MessageStrip>
-    );
-  }
-
-  if (containers.length === 1) {
-    return (
-      <SingleContainerSection
-        container={containers[0]}
-        containers={containers}
-        setContainers={setContainers}
-      />
-    );
-  }
-
-  return containers.map((container, i) => (
-    <ResourceForm.CollapsibleSection
-      key={i}
-      title={'Container ' + (container.name || i + 1)}
-      actions={
-        <Button
-          glyph="delete"
-          type="negative"
-          compact
-          onClick={() => removeContainer(i)}
-        />
-      }
-    >
-      <SingleContainerSection
-        container={container}
-        containers={containers}
-        setContainers={setContainers}
-      />
-    </ResourceForm.CollapsibleSection>
-  ));
 }
