+++ core-ui/src/components/Predefined/Create/Deployments/templates.js
@@ -49,51 +49,104 @@ export function createPresets(namespace, translate) {
   return [
     {
       name: translate('deployments.create-modal.presets.default'),
-      value: createDeploymentTemplate(namespace),
+      value: {
+        deployment: createDeploymentTemplate(namespace),
+      },
     },
     {
       name: 'Echo server',
       value: {
-        apiVersion: 'apps/v1',
-        kind: 'Deployment',
-        metadata: {
-          name: 'echo-server',
-          namespace,
-          labels: {},
-        },
-        spec: {
-          replicas: 1,
-          selector: {
-            matchLabels: {
-              app: 'echo-server',
-            },
+        deployment: {
+          apiVersion: 'apps/v1',
+          kind: 'Deployment',
+          metadata: {
+            name: 'echo-server',
+            namespace,
+            labels: {},
           },
-          template: {
-            metadata: {
-              labels: {
+          spec: {
+            replicas: 1,
+            selector: {
+              matchLabels: {
                 app: 'echo-server',
               },
             },
-            spec: {
-              containers: [
-                {
-                  name: 'echo-server',
-                  image: 'ealen/echo-server',
-                  resources: {
-                    requests: {
-                      memory: '64Mi',
-                      cpu: '50m',
-                    },
-                    limits: {
-                      memory: '128Mi',
-                      cpu: '100m',
+            template: {
+              metadata: {
+                labels: {
+                  app: 'echo-server',
+                },
+              },
+              spec: {
+                containers: [
+                  {
+                    name: 'echo-server',
+                    image: 'ealen/echo-server',
+                    resources: {
+                      requests: {
+                        memory: '64Mi',
+                        cpu: '50m',
+                      },
+                      limits: {
+                        memory: '128Mi',
+                        cpu: '100m',
+                      },
                     },
                   },
-                },
-              ],
+                ],
+              },
             },
           },
         },
+        service: createServiceTemplate(namespace),
+      },
+    },
+    {
+      name: 'HTTPBin',
+      value: {
+        deployment: {
+          apiVersion: 'apps/v1',
+          kind: 'Deployment',
+          metadata: {
+            name: 'httpbin',
+            namespace,
+            labels: {},
+          },
+          spec: {
+            replicas: 1,
+            selector: {
+              matchLabels: {
+                app: 'httpbin',
+              },
+            },
+            template: {
+              metadata: {
+                labels: {
+                  app: 'httpbin',
+                },
+              },
+              spec: {
+                containers: [
+                  {
+                    name: 'httpbin',
+                    image: 'kennethreitz/httpbin',
+                    resources: {
+                      requests: {
+                        memory: '64Mi',
+                        cpu: '50m',
+                      },
+                      limits: {
+                        memory: '128Mi',
+                        cpu: '100m',
+                      },
+                    },
+                  },
+                ],
+              },
+            },
+          },
+        },
+        service: createServiceTemplate(namespace),
       },
     },
   ];
